requirements.txt
asgiref==3.8.1
asttokens==2.4.1
boto3==1.35.4
botocore==1.35.4
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.6
django-factory-boy==1.0.0
django-filter==24.3
djangorestframework==3.15.2
exceptiongroup==1.2.1
executing==2.0.1
factory_boy==3.3.1
Faker==28.4.1
idna==3.7
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
jmespath==1.0.1
Markdown==3.7
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.2
sentry-sdk==2.13.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
